Renaming of app, add initdb command to create databases, docker-compose file env fixes
The app 'api' was renamed to resolver to better reflect what exactly the goal of this app is.  Flask/Cookiecutter does not seem to use the term 'app' to mean a generic Domain structure but instead to mean what the application actually does.  The app name 'api' was giving me path problems on ubuntu.
The command 'initdb' was added to allow for easier onboarding.  SQLAlchemy/alembic do not create databases when one is not found.  The user can now create those from the command line using SQLAlchemy-utils
The docker-compose was not able to pull in env variables due to a syntax error.
