--- api/auth/__init__.py
@@ -1,3 +1,0 @@
-from api.auth import views
-
-__all__ = ["views"]
