--- api/models/__init__.py
@@ -1,7 +1,0 @@
-from api.models.user import User
-from api.models.substance import Substance
-
-from api.models.blacklist import TokenBlacklist
-
-
-__all__ = ["Substance", "User", "TokenBlacklist"]
