--- api/resolver_api/__init__.py
@@ -1,3 +1,0 @@
-from api.resolver_api import views
-
-__all__ = ["views"]
