--- api/wsgi.py
@@ -1,3 +1,0 @@
-from api.app import create_app
-
-app = create_app()
