+++ requirements.txt
@@ -9,3 +9,4 @@ pytest-runner
 pytest-flask
 pytest-factoryboy
 factory_boy
+sqlalchemy-utils
