+++ resolver/api/__init__.py
@@ -1,0 +1,3 @@
+from resolver.api import views
+
+__all__ = ["views"]
