+++ resolver/app.py
@@ -1,14 +1,14 @@
 from flask import Flask
 
-from api import auth
-from api.resolver_api import views as api_views
-from api.extensions import db, jwt, migrate, apispec
+from resolver import auth
+from resolver.api import views as api_views
+from resolver.extensions import db, jwt, migrate, apispec, init_db
 
 
 def create_app(testing=False, cli=False):
     """Application factory, used to create application"""
-    app = Flask("api")
-    app.config.from_object("api.config")
+    app = Flask("resolver")
+    app.config.from_object("resolver.config")
     # this is the only place I was able to reliably set this variable
     # and avoid getting nagged about it
     app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
@@ -19,6 +19,8 @@ def create_app(testing=False, cli=False):
     configure_extensions(app, cli)
     configure_apispec(app)
     register_blueprints(app)
+
+    init_db()
 
     return app
