+++ resolver/auth/__init__.py
@@ -1,0 +1,3 @@
+from resolver.auth import views
+
+__all__ = ["views"]
