+++ resolver/commons/templates/swagger.j2
