+++ resolver/models/__init__.py
@@ -1,0 +1,7 @@
+from resolver.models.user import User
+from resolver.models.substance import Substance
+
+from resolver.models.blacklist import TokenBlacklist
+
+
+__all__ = ["Substance", "User", "TokenBlacklist"]
