+++ resolver/models/blacklist.py
@@ -14,7 +14,7 @@ like a redis or a memcached server.
 This example is heavily inspired by
 https://github.com/vimalloc/flask-jwt-extended/blob/master/examples/database_blacklist/
 """
-from api.extensions import db
+from resolver.extensions import db
 
 
 class TokenBlacklist(db.Model):
