+++ setup.py
@@ -3,7 +3,7 @@ from setuptools import setup, find_packages
 __version__ = "0.1"
 
 setup(
-    name="api",
+    name="resolver",
     version=__version__,
     packages=find_packages(exclude=["tests"]),
     install_requires=[
@@ -21,5 +21,5 @@ setup(
         "apispec[yaml]",
         "apispec-webframeworks",
     ],
-    entry_points={"console_scripts": ["api = api.manage:cli"]},
+    entry_points={"console_scripts": ["resolver = resolver.manage:cli"]},
 )
