+++ template.testenv
@@ -1,2 +1,2 @@
 SECRET_KEY=testing
-DATABASE_URI=postgresql://postgres:postgres@127.0.0.1:5431/resolver
+DATABASE_URI=postgresql://postgres:postgres@127.0.0.1:5431/test_resolver
