+++ tests/test_user.py
@@ -1,5 +1,5 @@
 from flask import url_for
-from api.models import User
+from resolver.models import User
 
 
 def test_get_user(client, db, user, admin_headers):
